+++ src/irmin-pack/unix/file_manager_intf.ml
@@ -65,6 +65,7 @@ module type S = sig
   module Index : Pack_index.S
   module Errs : Io_errors.S with module Io = Io
   module Sparse : Sparse_file.S with module Io = Io
+  module Lower : Lower.S with module Io = Io
 
   type t
 
@@ -75,6 +76,7 @@ module type S = sig
   val suffix : t -> Suffix.t
   val index : t -> Index.t
   val prefix : t -> Sparse.t option
+  val lower : t -> Lower.t option
 
   type create_error :=
     [ Io.create_error
@@ -82,7 +84,12 @@ module type S = sig
     | Io.open_error
     | Io.mkdir_error
     | `Corrupted_mapping_file of string
+    | `Corrupted_control_file
+    | `Double_close
+    | `Unknown_major_pack_version of string
+    | `Volume_missing of string
     | `Not_a_directory of string
+    | `Multiple_empty_volumes
     | `Index_failure of string ]
 
   val create_rw :
@@ -107,7 +114,7 @@ module type S = sig
     | `Index_failure of string
     | `Invalid_argument
     | `Invalid_layout
-    | `Io_misc of Control.Io.misc_error
+    | `Io_misc of Io.misc_error
     | `Migration_needed
     | `No_such_file_or_directory of string
     | `Not_a_directory of string
@@ -120,7 +127,8 @@ module type S = sig
     | `Unknown_major_pack_version of string
     | `Index_failure of string
     | `Sys_error of string
-    | `Inconsistent_store ]
+    | `Inconsistent_store
+    | `Volume_missing of string ]
 
   val open_rw : Irmin.Backend.Conf.t -> (t, [> open_rw_error ]) result
   (** Create a rw instance of [t] by opening existing files.
@@ -149,6 +157,7 @@ module type S = sig
     | `Migration_needed
     | `No_such_file_or_directory of string
     | `Not_a_file
+    | `Double_close
     | `Closed
     | `V3_store_from_the_future
     | `Index_failure of string
@@ -156,7 +165,8 @@ module type S = sig
     | `Inconsistent_store
     | `Invalid_argument
     | `Read_out_of_bounds
-    | `Invalid_layout ]
+    | `Invalid_layout
+    | `Volume_missing of string ]
 
   val open_ro : Irmin.Backend.Conf.t -> (t, [> open_ro_error ]) result
   (** Create a ro instance of [t] by opening existing files.
@@ -255,6 +265,7 @@ module type S = sig
   val generation : t -> int
   val gc_allowed : t -> bool
   val split : t -> (unit, [> Errs.t ]) result
+  val add_volume : t -> (unit, [> Errs.t ]) result
 
   val create_one_commit_store :
     t ->